refactor: moving utils files to resolve circular dependencies
